+++ src/hooks/useRandomBattleData/useRandomBattleData.ts
