+++ src/hooks/useWebsSocket/useWebsSocket.ts
@@ -1,5 +1,5 @@
 import React, { useEffect, useReducer, useRef, useState } from "react";
-import { isRandomBattle } from "../functions";
+import { isRandomBattle } from "../../functions";
 import {
   getActivePokemon,
   getBattleType,
